Moved <types.h> to <sys/types.h>. Removed dependencies on arch/x86/types.h - this file is not really used anymore, now. Might prevent compiling on Windows, though. Replaced "int" with "int32" for the id types. Removed some stuff from ktypes.h because it didn't belong there.
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@1120 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/private/kernel/arch/x86/types.h
@@ -13,8 +13,8 @@
   typedef __int64             int64;
 #endif
 
-typedef unsigned int      uint32;
-typedef int                int32;
+typedef unsigned long     uint32;
+typedef long               int32;
 typedef unsigned short    uint16;
 typedef short              int16;
 typedef unsigned char      uint8;
